Initial styling for the website
Adding some styling to the website.

Only have curated it for the home page so far but once I am happy with how the main page looks I will start styling the example project page and then use that as a template to create the project pages for my actual work.